Move Stream interface declaration to StreamManager
